Allow getting attribute in block
If an given address contains dots, it is ambiguous to find blocks from
the address without a schema.
To distinguish them in address notation requires introducing a strange new
syntax, which is not user friendly. The address notation is not specified
in the scope of the HCL specification, So the initial implementation has
Terraform in mind, but we want to solve it in a schemaless way.
We prioritize realistic usability over accuracy, we rely on some heuristics
here to compromise.
Given the address A.B.C, the user knows if B is a label or a nested block
type. So if the block matched in either, we should consider it is matched.
If you had both a label name and a nested block type, the address would
be A.B.B.C.
The labels take precedence over nested block. This is because if a block
type is specified, it is assumed that the number of labels in the same block
type does not really change and only the label name can be changed by the
user, and we want to give the user room to avoid unintended conflicts.