Add integration test
Adds `main_test.go` to test `main()` so that we can avoid pontential regression while implementing features like #4 that modifies code before executing *cobra.Command.

I was initially trying to build a PoC of #4 but then realized that the way I would implement it may result in breaking existing functionality in unexpected way. So please let me add a a bit more high-level test to ensure that it keeps the functionality at the `hcledit` command level, not function or cobra.Command level.