Add block append command
Allow us to append a new child block to matched blocks at a given parent
block address. The child address is relative to parent one.

We may want to insert it in the middle of body of the block instead of
at the end, in that case, we should implement it as a separate command
if necessary, because it will probably have different address format.

If a newline flag is true, it also appends a newline before the new
block.  Because the official hcl formatter doesn't have any opinion
about vertical whitespace, whether or not to add it is user's
preference.

In testing, the current mock implementation of newMockCmd doesn't
support parsing optional flags. It will require some refactorings to
support it. So, I temporary added newMockCmdWithFlag, but it should
merge to newMockCmd.