Add attribute append command
In #8, I added block append command. However it can add only an empty
body for simplicity. Of course, that's not make sense. In this patch, I
also added attribute append command, which is a missing parts for
appending a new block in a real use-case.

It allows us to append a new attribute to a given address.
If a matched block not found, nothing happens.
If the given attribute already exists, it returns an error.
If a newline flag is true, it also appends a newline before the new attribute.