Change verticalFormater to Filter instead of Sink
It was originally implemented as a Sink, but I found it's better as a Filter,
because using only default formater as a Sink is more simple and consistent.

As blockRemove has already been Filter, I added a new multiop Filter to
compose and apply multiple filters as a single Filter.