Abstract editor to Operator and split editing and deriving
The previous SinkEditor implementation was a wrong abstraction.
It was actually not editor.
This makes it hard to skip formatter in sink if no changes.

The hcledit provides not only operations for editing HCL,
but also for deriving such as listing.
They need similar but different implementations.

We introduce another abstraction: Operator, which is an interface which
abstracts stream operations.