Add support for getting nested block
Partial fix for #21

The previous implementation of the block get command didn't support a
nested block. However, it's natural to expect returning the nested
block as the same as the attribute in the nested block.

I'll reuse the findLongestMatchingBlocks to implement it as a compromise
for now, but it doesn't support the wildcard match. There is a bit
inconsistency here. To fix it, we will need to merge implementations of
findBlocks and findLongestMatchingBlocks.