Add body get command
Fixes #21

Add a new command `hcledit body get`, which gets body of first matched
block at a given address.

The current implementation doesn't support index in address format.
Merging body of contents for multiple blocks doesn't make sense,
so we take the first matched block.

Note: In the original feature request in #21, the expected behavior
contains the outer brackets `{}`, however, according to the HCL spec,
the outer brackets don't belong to the body.

https://github.com/hashicorp/hcl/blob/v2.9.1/hclsyntax/spec.md#structural-elements

So we should not include them in the outputs.