Skip formatter if filter didn't change contents to suppress meaningless diff
For replacing the tfupdate implementation using the hcledit,
we need to avoid meaningless diff because the tfupdate typically
processes a lot of files.

To implement this, I found the current implementation of EditOperator
implicitly assumes the Sink is a formatter.
It seems to be a bad abstraction, so I decide to distinguish them with
types.