Change input and output type for Operator.Apply() to []byte
In in-place update mode, it's better not to change the timestamp of file
if its contents has no change.

To implement this, we need to know the output contents changed. However
it's hard to compare input and output by stream.
In addition, the current implementation of HCL parser supports only bytes, not stream.
We cannot consume and apply filter in stream.
So, it's ok to apply filter in bytes.

I added new helper functions for using it in stream for convenience.