Add support for in-place update
Fixes #4

The hcledit was originally implemented as a stream editor to combine
other commands. However, I found it would be useful to update a given
file in-place in some cases. Of course, we can implement it by shell
redirection, it's not only boring, but also fragile for invalid input.
In addition, if we want to parse all files in a directory, we need to
write a script every time. It's not so fun.

This fix implements the in-place update for a single file. For a
directory, I'll add it later.

Initially I tried to imitate sed for the new arguments, but this wasn't
suitable. Because hcledit has different subcommands,  accepting
arbitrary variadic arguments is too restrictive for future extensions.
So I added a new global flag `-f (--file)` for an input file.

In addition, I'm thinking of replacing the hcl related implementation of
the tfupdate with the hcledit, the tfupdate has already used `-i` for
`--ignore-path`. Using `-i` for in-place update will probably confuse
the tfupdate users. So I chose `-u (--update)` for in-place update.