Add attribute mv command
I believe the pattern "attribute A is deprecated, use B instead” is a
generic use case. The current functionality only allows removing the old
attribute and appending a new one to the end, which changes the order of
the attributes. I don't want to change the order because meaningful
order contributes to reducing cognitive load. To help with such a
transition, it would be great if we could rename an attribute.

Regarding the design, we have aligned the attribute mv command with
block mv but have not implemented moving across blocks yet. However, it
accepts an address as an argument so that it can be extended in the
future.

As for implementation, hclwrite's attribute name is currently a private
variable with no API for renaming it. Therefore, we will use the forked
version until the patch is merged.