Change the argument of the block new command to accept an address
The current interface of the block new command is inconsistent when
compared to the block append command in terms of passing labels for
blocks to be added, space-delimited, and dot-delimited. The
dot-delimited address notation is ambiguous concerning nested blocks.
Still, even if someone wanted to create nested blocks, space
delimitation would not solve the problem, and neither is sufficiently
expressive. Therefore, I have decided that it is more consistent and
reasonable to accept the limitation that nested blocks cannot be created
in a single shot, align the notation with block append, and change how
the block new command takes arguments.