Add attribute replace command
In addition to the attribute mv command, add a new attribute replace
command. This command not only renames the key of an attribute but also
sets the value of the attribute. This could be done by combining
attribute mv and attribute set, but doing it with a single command would
be more convenient. This is because in dependency upgrade scenarios,
where attribute A is deprecated and use B instead, it is often the case
that not only the name of the attribute changes but also its value.

For example, Terraform v1.10 introduced DynamoDB-free S3-native state
locking in the s3 backend, and the upcoming Terraform v1.11 will start
deprecating the old dynamodb_table attribute and recommend using the new
use_lockfile. It's easy to rewrite a few files by hand, but it could be
tedious if you have hundreds of backend configurations across many
repositories.

While the above is the initial motivation for implementing this feature,
it is not hard to imagine that the real-world use cases go beyond this.

As with the attribute mv command, a forked version of hclwrite is used
to implement this feature. So, we need to wait for the upstream patch to
be merged before we merge this implementation.